requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.2.0
